Enumeration Update:
Researched Enumeration yesterday and decided to try it out as an option
to importing the data for the grocery list.

I plan to limit the grocery list items to 10 items per category for
simplicity's sake.

I have started writing a few Enum classes and displayed sample
functionality to show how enums can be used for our project idea,
hopefully.

Please run the program from the ListInput file to see the relevant
output in the console.