I have created the remaining Enum classes to each hold one Item Category such as Dairy, Meats, etc. I have also finalized the Options applicable to each enum class.
Since Organic is the only Option that most grocery items have in common,
I might create an Interface for the Organic Option so that we can show
proof of that learning outcome.

I am also thinking about what classes to extend to show the LO for
inheritance and polymorphism. I'm thinking that for the VIEW we could
extend the JFrame class and/or the JButton class depending on which
option will be easier to implement.

We have to use some type of data structure. Since we are using Arrays
for the enum constants, maybe we can use a Queue or a LinkedList to
store the items that the user adds to the list. Queue would be a good
option because it is FIFO and that way the exported list would be in the
order the user added items to it. However, LinkedList might also be a
good choice because the user can easily retrieve an item's location and
add or remove items to and from the list easily, which we might want the
user to be able to do.

For File I/O LO, I thought we could demonstrate File Input by importing
the Food Quotes List, and to demonstrate File Output, we could export
the finished grocery list.

For Exception Handling LO, we could simply write exceptions to catch
whenever a user might want to select an option for an item that doesn't
exist, and/or if the user attempts to export an empty list. An exception
could be thrown, and the user could receive an error message.